Crack artifact problem solved. there is some mistakes what i made. 1. inform to gpu that data type is unsigned integer but actually i insert float data. 2. in shader level, mix(a, b, c) and mix(b, a, c) are very different.

+++ Renderer/resources/shader/terrain_vs.glsl
@@ -1,9 +1,9 @@
 #version 430 core
 
 layout(location = 0) in vec3 aPos;
-layout(location = 1) in uint aTessLevel;
+layout(location = 1) in float aTessLevel;
 
-out uint vs_tessLevel;
+out float vs_tessLevel;
 out vec2 vs_texCoords;
 
 uniform vec3 originPos;
